1.0/requirements.txt
openai==0.28.0
biopython==1.83
pandas==2.2.1
openpyxl==3.1.2
requests==2.31.0